# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Suparman API
  description: |
    Suparman is a database synchronization and keep-alive platform for Supabase.
    This API allows you to manage database connections, sync jobs, and monitor system health.
  version: 1.0.0
  contact:
    name: Suparman Support
    url: https://suparbase.com
  license:
    name: MIT

servers:
  - url: https://suparbase.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server

tags:
  - name: Health
    description: Health check endpoints
  - name: Connections
    description: Database connection management
  - name: Sync
    description: Sync job operations
  - name: Explorer
    description: Database explorer
  - name: Admin
    description: Admin operations (requires admin privileges)

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns detailed health status of all system components
      operationId: getHealth
      responses:
        '200':
          description: System is healthy or degraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: System is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /backend-health:
    get:
      tags:
        - Health
      summary: Backend health check
      description: Returns health status of the backend server
      operationId: getBackendHealth
      responses:
        '200':
          description: Backend health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendHealthResponse'

  /connections:
    get:
      tags:
        - Connections
      summary: List connections
      description: Returns all database connections for the authenticated user
      operationId: listConnections
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of connections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Connections
      summary: Create connection
      description: Creates a new database connection
      operationId: createConnection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionInput'
      responses:
        '201':
          description: Connection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /connections/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Connections
      summary: Get connection
      description: Returns a specific connection by ID
      operationId: getConnection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Connections
      summary: Update connection
      description: Updates an existing connection
      operationId: updateConnection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionUpdateInput'
      responses:
        '200':
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Connections
      summary: Delete connection
      description: Deletes a connection
      operationId: deleteConnection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connection deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /connections/{id}/test:
    post:
      tags:
        - Connections
      summary: Test connection
      description: Tests if a connection is working
      operationId: testConnection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTestResponse'

  /connections/{id}/schema:
    get:
      tags:
        - Connections
      summary: Get database schema
      description: Returns the database schema for a connection
      operationId: getConnectionSchema
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Database schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaResponse'

  /sync:
    get:
      tags:
        - Sync
      summary: List sync jobs
      description: Returns all sync jobs for the authenticated user
      operationId: listSyncJobs
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of sync jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobsResponse'

    post:
      tags:
        - Sync
      summary: Create sync job
      description: Creates a new sync job
      operationId: createSyncJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncJobInput'
      responses:
        '201':
          description: Sync job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobResponse'

  /sync/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Sync
      summary: Get sync job
      description: Returns a specific sync job
      operationId: getSyncJob
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sync job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobResponse'

    delete:
      tags:
        - Sync
      summary: Delete sync job
      description: Deletes a sync job
      operationId: deleteSyncJob
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sync job deleted

  /sync/{id}/start:
    post:
      tags:
        - Sync
      summary: Start sync job
      description: Starts or resumes a sync job
      operationId: startSyncJob
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sync job started

  /sync/{id}/stop:
    post:
      tags:
        - Sync
      summary: Stop sync job
      description: Stops a running sync job
      operationId: stopSyncJob
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sync job stopped

  /sync/{id}/pause:
    post:
      tags:
        - Sync
      summary: Pause sync job
      description: Pauses a running sync job
      operationId: pauseSyncJob
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sync job paused

  /explorer/{connectionId}/tables:
    get:
      tags:
        - Explorer
      summary: List tables
      description: Lists all tables in a database
      operationId: listTables
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TablesResponse'

  /explorer/{connectionId}/{table}/rows:
    get:
      tags:
        - Explorer
      summary: Get table rows
      description: Returns paginated rows from a table
      operationId: getTableRows
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: orderBy
          in: query
          schema:
            type: string
            default: id
        - name: orderDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Table rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowsResponse'

  /admin/users:
    get:
      tags:
        - Admin
      summary: List users
      description: Lists all users (admin only)
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
        '403':
          $ref: '#/components/responses/Forbidden'

  /admin/sync-jobs:
    get:
      tags:
        - Admin
      summary: List all sync jobs
      description: Lists all sync jobs across all users (admin only)
      operationId: listAllSyncJobs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all sync jobs
        '403':
          $ref: '#/components/responses/Forbidden'

  /admin/analytics:
    get:
      tags:
        - Admin
      summary: Get analytics
      description: Returns system analytics (admin only)
      operationId: getAnalytics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Analytics data
        '403':
          $ref: '#/components/responses/Forbidden'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        uptime:
          type: integer
          description: Uptime in seconds
        checks:
          type: object
          properties:
            supabase:
              $ref: '#/components/schemas/HealthCheck'
            backend:
              $ref: '#/components/schemas/HealthCheck'
            redis:
              $ref: '#/components/schemas/HealthCheck'
            encryption:
              $ref: '#/components/schemas/HealthCheck'
        metrics:
          type: object
          properties:
            memory:
              type: object
              properties:
                used:
                  type: integer
                total:
                  type: integer
                percentage:
                  type: integer

    HealthCheck:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error, degraded, unknown]
        latency:
          type: integer
        message:
          type: string

    BackendHealthResponse:
      type: object
      properties:
        healthy:
          type: boolean
        status:
          type: string
        latency:
          type: integer

    ConnectionInput:
      type: object
      required:
        - name
        - databaseUrl
        - environment
      properties:
        name:
          type: string
          maxLength: 100
        databaseUrl:
          type: string
          format: uri
          description: PostgreSQL connection URL
        environment:
          type: string
          enum: [production, development]

    ConnectionUpdateInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        environment:
          type: string
          enum: [production, development]

    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        environment:
          type: string
          enum: [production, development]
        keep_alive:
          type: boolean
        last_pinged_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    ConnectionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Connection'

    ConnectionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Connection'

    ConnectionTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            connected:
              type: boolean
            latency:
              type: integer
            version:
              type: string

    SchemaResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            tables:
              type: array
              items:
                type: object

    SyncJobInput:
      type: object
      required:
        - sourceConnectionId
        - targetConnectionId
        - direction
        - tables
      properties:
        sourceConnectionId:
          type: string
          format: uuid
        targetConnectionId:
          type: string
          format: uuid
        direction:
          type: string
          enum: [one_way, two_way]
        tables:
          type: array
          items:
            type: object
            properties:
              tableName:
                type: string
              enabled:
                type: boolean
              conflictStrategy:
                type: string
                enum: [last_write_wins, source_wins, target_wins, manual]
        dryRun:
          type: boolean
          default: false

    SyncJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_connection_id:
          type: string
          format: uuid
        target_connection_id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [one_way, two_way]
        status:
          type: string
          enum: [pending, running, completed, failed, paused]
        progress:
          type: object
        created_at:
          type: string
          format: date-time

    SyncJobsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/SyncJob'
        pagination:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer

    SyncJobResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/SyncJob'

    TablesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            tables:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  rowCount:
                    type: integer
                  columnCount:
                    type: integer
                  columns:
                    type: array
                    items:
                      type: object
            total:
              type: integer

    RowsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                type: object
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            hasMore:
              type: boolean